requirements.txt
asgiref==3.4.1
certifi==2021.10.8
chardet==3.0.4
Django==3.2.8
django-cors-headers==3.10.0
django-extra-fields==3.0.2
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
logging-logger==0.1
mysqlclient==2.0.3
Pillow==8.4.0
psycopg2==2.9.1
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
rfc3986==1.5.0
sniffio==1.2.0
sqlparse==0.4.2
